Simplify design
