+++ src/Objectivity.AutoFixture.XUnit2.Core.Tests/Attributes/CustomizeWithAttributeTests.cs
@@ -10,6 +10,7 @@
     using global::AutoFixture;
     using global::AutoFixture.Kernel;
     using global::AutoFixture.Xunit2;
+
     using Objectivity.AutoFixture.XUnit2.Core.Attributes;
     using Objectivity.AutoFixture.XUnit2.Core.Customizations;
 
@@ -92,16 +93,20 @@
             Assert.Throws<ArgumentNullException>(() => new CustomizeWithAttribute(customizationType));
         }
 
-        [Fact(DisplayName = "GIVEN unsupported type WHEN constructor is invoked THEN exception is thrown")]
-        public void GivenUnsupportedType_WhenConstructorIsInvoked_ThenExceptionIsThrown()
+        [Fact(DisplayName = "GIVEN unsupported type WHEN GetCustomization is invoked THEN exception is thrown")]
+        public void GivenUnsupportedType_WhenGetCustomizationIsInvoked_ThenExceptionIsThrown()
         {
             // Arrange
             var customizationType = typeof(string);
+            var customizeAttribute = new CustomizeWithAttribute(customizationType);
+            var parameter = typeof(CustomizeWithAttributeTests)
+                .GetMethod(nameof(this.MethodUnderTest), BindingFlags.Instance | BindingFlags.NonPublic)
+                .GetParameters()
+                .First();
 
             // Act
             // Assert
-            var exception = Assert.Throws<ArgumentException>(() => new CustomizeWithAttribute(customizationType));
-            exception.Message.Should().NotBeNullOrEmpty().And.Contain(nameof(ICustomization));
+            Assert.Throws<InvalidOperationException>(() => customizeAttribute.GetCustomization(parameter));
         }
 
         [Fact(DisplayName = "GIVEN CustomizeWith attribute with IncludeParameterType set WHEN GetCustomization is invoked THEN customization with expected type is returned")]
