+++ src/Objectivity.AutoFixture.XUnit2.Core/CustomisationFactories/ICustomisationFactoryProvider.cs
@@ -1,12 +1,9 @@
 ﻿namespace Objectivity.AutoFixture.XUnit2.Core.CustomisationFactories
 {
     using System;
-    using System.Collections.ObjectModel;
 
     internal interface ICustomisationFactoryProvider
     {
-        ReadOnlyCollection<Type> SupportedTypes { get; }
-
         ICustomisationFactory GetFactory(Type type);
     }
 }
